Add files via upload
In the context of daily share trading, a share trader buys and sells shares within the same day. The trader is restricted to a maximum of 2 transactions in a day, with the condition that the second transaction can only commence after the completion of the first one (Buy -> Sell -> Buy -> Sell).